# src/domains/products/schemas.py
from pydantic import BaseModel, Field

class ProductBase(BaseModel):
    name: str
    price: float = Field(gt=0)
    stock: int = Field(ge=0)
    category: str

class ProductCreate(ProductBase):
    pass

class ProductResponse(ProductBase):
    id: str